Pull out TableOfContentsItems
